Begin implementing servo control algorithm
There is now a timer which 'plays back' a set of pin-state/duration
commands. The execution time from the start of the interrupt to the next
duration being set is constant, although I've now realised this is not
critical as the timer will still be counting from the previous cycle.
This means that the long-term timing will be perfect, however there
could be 1-cycle of jitter on the individual pulse timings due to the
interrupt latency jitter. This will be insignificant in the grand scheme
of things.

Still to be implemented is the function which fills in the set of
commands, based upon the current servo position settings. To begin with
I'm going to make this really simple and just output the pulses in the
sequence 0-11. Once that is working I will then implement the sorting of
pulse lengths so that it actually works when you have an arbitrary
combination of pulse lengths.